requirements.txt
Flask==2.3.2
requests==2.31.0
firebase-admin==6.4.0
google-cloud-firestore==2.13.0
gTTS==2.5.1
python-dotenv==1.0.1
